huge refactor- split code into multiple files, use readFile instead of stdin
