+++ __tests__/fixtures/test.csv
@@ -1,0 +1,3 @@
+Timestamp,My GitHub ID is:,"Do you participate in open source communities that are not on GitHub? If so, please tell us more!",,Email Address
+9/5/2018 16:46:48,fakeUser1,,,firstFake@indeed.com
+9/5/2018 17:26:29,fakeUser2,ASF ID: second,,secondFake@indeed.com
